+++ tests/boot/with_mcumgr/pytest/test_upgrade.py
@@ -8,14 +8,13 @@ import logging
 
 from pathlib import Path
 from twister_harness import DeviceAdapter, Shell, MCUmgr
-from west_sign_wrapper import west_sign_with_imgtool
-from utils import (
+from twister_harness.helpers.utils import (
     find_in_config,
     match_lines,
-    match_no_lines,
-    check_with_shell_command,
-    check_with_mcumgr_command,
+    match_no_lines
 )
+from utils import check_with_shell_command, check_with_mcumgr_command
+from west_sign_wrapper import west_sign_with_imgtool
 
 logger = logging.getLogger(__name__)
